DistributionUtils

CalibrateEDMF.DistributionUtils.construct_priors — Function

construct_priors(
    params::Dict{String, Vector{Constraint}};
    unconstrained_σ::Float64 = 1.0,
    prior_mean::Union{Dict{String, Vector{Float64}}, Nothing} = nothing,
    outdir_path::String = pwd(),
    to_file::Bool = true,
)

Define a prior Gaussian ParameterDistribution in unconstrained space from a dictionary of constraints.

This constructor assumes independent priors and the same unconstrained standard deviation for each parameter. Note that the standard deviation in unconstrained space is normalized with respect to the constrained interval width, so it automatically takes into account parameter scales.

The constructor also allows passing a prior mean for each parameter in constrained space.

Inputs:

params :: Dictionary of parameter names to constraints.

unconstrained_σ :: Standard deviation of the transformed gaussians (unconstrained space).

prior_mean :: The mean value of the prior in constrained space. If not given, the prior is selected to be 0 in unconstrained space.

outdir_path :: Output path.

to_file :: Whether to write the serialized prior to a JLD2 file.

Output:

The prior ParameterDistribution.

CalibrateEDMF.DistributionUtils.deserialize_prior — Function

deserialize_prior(prior_dict::Dict{String, Any})

Generates a prior ParameterDistribution from arguments stored in a dictionary.

CalibrateEDMF.DistributionUtils.flatten_config_dict — Function

flatten_config_dict(param_dict::Dict{String, Vector{T}})

Given a dictionary of parameter names to parameter vectors of arbitrary length, return a new dictionary that maps a unique parameter name to each element of the full flattened vector of parameters.

Inputs:

param_dict :: Dictionary of parameter names to constraints.

Outputs:

u_names :: Vector{String} :: vector of parameter names

values :: Vector{Vector{T}} :: vector of single-valued vectors encapsulating parameter values.

CalibrateEDMF.DistributionUtils.flat_dict_keys_where — Function

flat_dict_keys_where(dict::Dict{String, Vector{T}}, condition::Function = identity) where {T}

Flattens the values of a dictionary with parameter vectors as keys, and returns the indices of entries in the flattened dictionary satisfying a given condition.

Inputs: param_dict :: Dictionary of parameter names to vectors. condition :: A condition function operating on each dictionary value. Outputs: Indices of flattened entries satisfying the condition.

CalibrateEDMF.DistributionUtils.logmean_and_logstd — Function

logmean_and_logstd(μ, σ)

Returns the lognormal parameters μ and σ from the mean μ and std σ of the lognormal distribution.

CalibrateEDMF.DistributionUtils.mean_and_std_from_ln — Function

mean_and_std_from_ln(μ, σ)

Returns the mean and variance of the lognormal distribution from the lognormal parameters μ and σ.