LESUtils

CalibrateEDMF.LESUtils.get_les_names — Function

get_les_names(y_names::Vector{String}, filename::String)
get_les_names(m::ReferenceModel, filename::String)

Returns the aliases of the variables actually present in the nc file (filename) corresponding to SCM variables y_names.

CalibrateEDMF.LESUtils.find_alias — Function

find_alias(aliases::Tuple{Vararg{String}}, filename::String)

Finds the alias present in an NCDataset from a list of possible aliases.

CalibrateEDMF.LESUtils.get_cfsite_les_dir — Function

get_cfsite_les_dir(
    cfsite_number::Integer;
    forcing_model::String = "HadGEM2-A",
    month::Integer = 7,
    experiment::String = "amip",)

Given information about an LES run from Zhaoyi Shen, Akshay Sridhar, Zhihong Tan, Anna Jaruga, Tapio Schneider (2022), fetch LES directory on central cluster.

Inputs:

cfsite_number :: cfsite number

forcing_model :: {"HadGEM2-A", "CNRM-CM5", "CNRM-CM6-1", "IPSL-CM6A-LR"} - name of climate model used for forcing. Currently, only "HadGEM2-A" simulations are available reliably.

month :: {1, 4, 7, 10} - month of simulation.

experiment :: {"amip", "amip4K"} - experiment from which LES was forced.

Outputs:

les_dir - path to les simulation containing stats folder

CalibrateEDMF.LESUtils.get_LES_library — Function

get_LES_library

Hierarchical dictionary of available LES simulations described in Zhaoyi Shen, Akshay Sridhar, Zhihong Tan, Anna Jaruga, Tapio Schneider (2022). The following cfsites are available across listed models, months, and experiments.