ReferenceModels

CalibrateEDMF.ReferenceModels.ReferenceModel — Type

ReferenceModel

A structure containing information about the 'true' reference model and the observation map used to compare the parameterized and reference models.

Fields

y_names::Vector{String}

Vector of reference variable names

y_dir::String

Directory for reference data to compute y mean vector

Σ_dir::String

Directory for reference data to compute Σ covariance matrix

scm_dir::String

Directory for static data related to forward scm model (parameter file & vertical levels)

case_name::String

Name of case

y_t_start::Real

Start time for computing mean statistics over

y_t_end::Real

End time for computing mean statistics over

Σ_t_start::Real

Start time for computing covariance statistics over

Σ_t_end::Real

End time for computing covariance statistics over

n_obs::Union{Nothing, Integer}

Number of observed vertical locations, if not the native grid

CalibrateEDMF.ReferenceModels.construct_reference_models — Function

construct_reference_models(kwarg_ld::Dict{Symbol, Vector{T} where T})::Vector{ReferenceModel}

Returns a vector of ReferenceModels given a dictionary of keyword argument lists.

Inputs:

kwarg_ld :: Dictionary of keyword argument lists

Outputs:

ref_models :: Vector where the i-th ReferenceModel is constructed from the i-th element of every keyword argument list of the dictionary.

CalibrateEDMF.ReferenceModels.time_shift_reference_model — Function

time_shift_reference_model(m::ReferenceModel, Δt::FT) where {FT <: Real}

Returns a time-shifted ReferenceModel, considering an interval relative to the last available time step of the original model.

Inputs:

m :: A ReferenceModel.

Δt :: [LES last time - SCM start time (LES timeframe)]

Outputs:

The time-shifted ReferenceModel.

CalibrateEDMF.ReferenceModels.get_z_obs — Function

Returns the observed vertical locations for a reference model

CalibrateEDMF.ReferenceModels.ReferenceModelBatch — Type

struct ReferenceModelBatch

A structure containing a batch of ReferenceModels and a mutable evaluation order for ReferenceModels within the current epoch.

CalibrateEDMF.ReferenceModels.get_minibatch! — Function

get_minibatch!(ref_models::ReferenceModelBatch, batch_size::Int)

Returns a minibatch of ReferenceModels from a ReferenceModelBatch and updates the eval order.

The size of the minibatch is either the requested size, or the remainder of the elements in the eval_order for this epoch.

Inputs:

ref_model_batch :: A ReferenceModelBatch.

batch_size :: The number of ReferenceModels to retrieve.

Outputs:

A vector of ReferenceModels.

The indices of the returned ReferenceModels.

CalibrateEDMF.ReferenceModels.reshuffle_on_epoch_end — Function

Restarts a shuffled evaluation order if the current epoch has finished.