ReferenceStats

CalibrateEDMF.ReferenceStats.ReferenceStatistics — Type

ReferenceStatistics{FT <: Real}

A structure containing statistics from the reference model used to define a well-posed inverse problem.

Fields

y::Vector{FT} where FT<:Real

Reference data, length: nSim * n_vars * n_zLevels (possibly reduced by PCA)

Γ::Matrix{FT} where FT<:Real

Data covariance matrix, dims: (y,y) (possibly reduced by PCA)

norm_vec::Array{Vector{FT}, 1} where FT<:Real

Vector (length: nSim) of normalizing factors (length: n_vars)

pca_vec::Array{Union{LinearAlgebra.UniformScaling, Matrix{FT}}, 1} where FT<:Real

Vector (length: nSim) of PCA projection matrices with leading eigenvectors as columns

y_full::Vector{FT} where FT<:Real

Full reference data vector, length: nSim * n_vars * n_zLevels

Γ_full::SparseArrays.SparseMatrixCSC{FT, Int64} where FT<:Real

Full covariance matrix, dims: (y,y)

CalibrateEDMF.ReferenceStats.get_obs — Function

get_obs(
    m::ReferenceModel,
    y_names::Vector{String},
    Σ_names::Vector{String},
    normalize::Bool;
    z_scm::Union{Vector{FT}, Nothing} = nothing,
)

Get observations for variables ynames, interpolated to zscm (if given), and possibly normalized with respect to the pooled variance.

Inputs:

obs_type :: Either :les or :scm

m :: Reference model

z_scm :: If given, interpolate LES observations to given levels.

Outputs:

y :: Mean of observations, possibly interpolated to z_scm levels.

Σ :: Observational covariance matrix, possibly pool-normalized.

pool_var :: Vector of vertically averaged time-variance, one entry for each variable

CalibrateEDMF.ReferenceStats.obs_PCA — Function

obs_PCA(y_mean, y_var, allowed_var_loss = 1.0e-1)

Perform dimensionality reduction using principal component analysis on the variance yvar. Only eigenvectors with eigenvalues that contribute to the leading 1-allowedvar_loss variance are retained. Inputs:

y_mean :: Mean of the observations.

y_var :: Variance of the observations.

allowed_var_loss :: Maximum variance loss allowed.

Outputs:

y_pca :: Projection of y_mean onto principal subspace spanned by eigenvectors.

y_var_pca :: Projection of y_var on principal subspace.

P_pca :: Projection matrix onto principal subspace, with leading eigenvectors as columns.

CalibrateEDMF.ReferenceStats.pca — Function

pca(covmat::AbstractMatrix{FT}, allowed_var_loss::FT) where {FT <: Real}

Perform dimensionality reduction using principal component analysis on the variance covmat.

Inputs:

covmat :: Variance of the observations.

allowed_var_loss :: Maximum variance loss allowed.

Outputs:

λ_pca :: Principal eigenvalues, ordered in increasing value order.

P_pca :: Projection matrix onto principal subspace, with leading eigenvectors as columns.

CalibrateEDMF.ReferenceStats.get_profile — Function

get_profile(
    filename::String,
    var_names::Vector{String};
    ti::Real = 0.0,
    tf = nothing,
    z_scm::Union{Vector{FT}, Nothing} = nothing,
)

Get profiles for variables varnames, interpolated to zscm (if given), and concatenated into a single output vector.

Inputs:

filename :: nc filename

var_names :: Names of variables to be retrieved.

z_scm :: If given, interpolate LES observations to given levels.

Outputs:

y :: Output vector used in the inverse problem, which concatenates the requested profiles.

CalibrateEDMF.ReferenceStats.get_time_covariance — Function

get_time_covariance(m::ReferenceModel, var_names::Vector{String}; z_scm::Vector{FT}) where {FT <: Real}

Obtain the covariance matrix of a group of profiles, where the covariance is obtained in time. Inputs:

m :: Reference model.

var_names :: List of variable names to be included.

z_scm :: If given, interpolates covariance matrix to this locations.