TurbulenceConvectionUtils

CalibrateEDMF.TurbulenceConvectionUtils.ModelEvaluator — Type

struct ModelEvaluator

A structure containing the information required to perform a forward model evaluation at a given parameter vector.

CalibrateEDMF.TurbulenceConvectionUtils.run_SCM — Function

run_SCM(
    u::Vector{FT},
    u_names::Vector{String},
    RM::Vector{ReferenceModel},
    RS::ReferenceStatistics;
    error_check::Bool = false,
    namelist_args = nothing,
    failure_handler = "high_loss",
) where {FT <: Real}

Run the single-column model (SCM) using a set of parameters u and return the value of outputs defined in y_names, possibly after normalization and projection onto lower dimensional space using PCA.

Inputs:

u :: Values of parameters to be used in simulations.

u_names :: SCM names for parameters u.

RM :: Vector of ReferenceModels

RS :: reference statistics for simulation

error_check :: Returns as an additional argument whether the SCM call errored.

namelist_args :: Additional arguments passed to the TurbulenceConvection namelist.

failure_handler :: Method used to handle failed simulations.

Outputs:

sim_dirs :: Vector of simulation output directories

g_scm :: Vector of model evaluations concatenated for all flow configurations.

g_scm_pca :: Projection of g_scm onto principal subspace spanned by eigenvectors.

model_error :: Whether the simulation errored with the requested configuration.

CalibrateEDMF.TurbulenceConvectionUtils.run_reference_SCM — Function

run_reference_SCM(m::ReferenceModel; overwrite::Bool = false)

Run the single-column model (SCM) for a reference model object using default parameters.

Inputs:

m :: A ReferenceModel.

overwrite :: if true, overwrite existing simulation files.

run_single_timestep :: if true, run only one time step.

namelist_args :: Additional arguments passed to the TurbulenceConvection namelist.

CalibrateEDMF.TurbulenceConvectionUtils.eval_single_ref_model — Function

eval_single_ref_model(
    m_index::IT,
    m::ReferenceModel,
    RS::ReferenceStatistics,
    u::Vector{FT},
    u_names::Vector{String},
    namelist_args = nothing,
) where {FT <: Real, IT <: Int}

Runs the single-column model (SCM) under a single configuration (i.e., ReferenceModel) using a set of parameters u, and returns the forward model evaluation in both the original and the latent PCA space.

Inputs:

m_index :: The index of the ReferenceModel within the overarching ref_models vector used to construct the ReferenceStatistics.

m :: A ReferenceModel.

RS :: reference statistics for simulation

u :: Values of parameters to be used in simulations.

u_names :: SCM names for parameters u.

namelist_args :: Additional arguments passed to the TurbulenceConvection namelist.

Outputs:

sim_dir :: Simulation output directory.

g_scm :: Forward model evaluation in original output space.

g_scm_pca :: Projection of g_scm onto principal subspace spanned by eigenvectors.

model_error :: Whether the simulation errored with the requested configuration.

CalibrateEDMF.TurbulenceConvectionUtils.run_SCM_handler — Function

run_SCM_handler(
    m::ReferenceModel,
    tmpdir::String,
    u::Array{FT, 1},
    u_names::Array{String, 1},
    namelist_args = nothing,
) where {FT<:AbstractFloat}

Run a case using a set of parameters u_names with values u, and return directory pointing to where data is stored for simulation run.

Inputs:

m :: Reference model

tmpdir :: Temporary directory to store simulation results in

u :: Values of parameters to be used in simulations.

u_names :: SCM names for parameters u.

namelist_args :: Additional arguments passed to the TurbulenceConvection namelist.

Outputs:

output_dir :: directory containing output data from the SCM run.

model_error :: Boolean specifying whether the simulation failed.

run_SCM_handler(
    case_name::String,
    out_dir::String;
    u::Vector{FT},
    u_names::Vector{String},
    namelist::Union{Dict, Nothing} = nothing,
    namelist_args::Union{Tuple, Nothing} = nothing,
    uuid::String = "01",
    les::Union{NamedTuple, String} = nothing,
)

Run a TurbulenceConvection.jl case and return directory pointing to where data is stored for simulation run.

Inputs:

case_name :: case name

out_dir :: Directory to store simulation results in.

Optional Inputs:

u :: Values of parameters to be used in simulations.

u_names :: SCM names for parameters u.

namelist :: namelist to use for simulation.

namelist_args :: Additional arguments passed to the TurbulenceConvection namelist.

uuid :: uuid of SCM run

les :: path to LES stats file, or NamedTuple with keywords {forcingmodel, month, experiment, cfsitenumber} needed to specify path.

Outputs:

output_dir :: directory containing output data from the SCM run.

model_error :: Boolean specifying whether the simulation failed.

CalibrateEDMF.TurbulenceConvectionUtils.get_scm_namelist — Function

get_scm_namelist(m::ReferenceModel; overwrite::Bool = false)::Dict

Fetch the namelist stored in scm_dir(m). Generate a new namelist if it doesn't exist or overwrite=true.

CalibrateEDMF.TurbulenceConvectionUtils.create_parameter_vectors — Function

create_parameter_vectors(u_names::Vector{String}, u::Vector{FT}) where {FT <: AbstractFloat}

Given vector of parameter names and corresponding values, combine any vector components into single parameter vectors for input into SCM.

Inputs:

u_names :: SCM names for parameters u, which may contain vector components.

u :: Values of parameters to be used in simulations, which may contain vector components.

Outputs:

u_names_out :: SCM names for parameters u.

u_out :: Values of parameters to be used in simulations.

CalibrateEDMF.TurbulenceConvectionUtils.generate_scm_input — Function

generate_scm_input(
    model_evaluators::Vector{ModelEvaluator{FT}},
    outdir_path::String = pwd(),
) where {FT <: AbstractFloat}

Writes to file a set of ModelEvaluator used to initialize SCM evaluations at different parameter vectors, as well as their assigned numerical version.

CalibrateEDMF.TurbulenceConvectionUtils.get_gcm_les_uuid — Function

get_gcm_les_uuid(
    cfsite_number::Integer;
    forcing_model::String,
    month::Integer,
    experiment::String,)

Generate unique and self-describing uuid given information about a GCM-driven LES simulation from Shen et al. 2022.

CalibrateEDMF.TurbulenceConvectionUtils.precondition — Function

precondition(
    param::Vector{FT},
    priors,
    ref_models::Vector{ReferenceModel},
    ref_stats::ReferenceStatistics,
    namelist_args = nothing;
    counter::Integer = 0,
    max_counter::Integer = 10,
) where {FT <: Real}

Substitute parameter vector param by a parameter vector drawn from the same prior, conditioned on the forward model being stable.

Inputs:

param :: A parameter vector that may possibly result in unstable forward model evaluations (in unconstrained space).

priors :: Priors from which the parameters were drawn.

ref_models :: Vector of ReferenceModels to check stability for.

ref_stats :: ReferenceStatistics of the ReferenceModels.

namelist_args :: Arguments passed to the TC.jl namelist.

counter :: Accumulator tracking number of recursive calls to preconditioner.

max_counter :: Maximum number of recursive calls to the preconditioner.

Outputs:

new_param :: A new parameter vector drawn from the prior, conditioned on

simulations being stable (in unconstrained space).

precondition(ME::ModelEvaluator, priors)

Substitute the parameter vector of a ModelEvaluator by another one drawn from the given priors, conditioned on the forward model being stable.

Inputs:

ME :: A ModelEvaluator.

priors :: Priors from which the parameters were drawn.

Outputs:

A preconditioned ModelEvaluator.