Diagnostics

CalibrateEDMF.Diagnostics.io_dictionary_metrics — Function

io_dictionary_metrics()

Scalar metrics dictionary.

Elements:

loss_mean_g :: (ḡ - y)'Γ_inv(ḡ - y).

mse_full_mean :: Ensemble mean of MSE(g_full, y_full).

mse_full_min :: Ensemble min of MSE(g_full, y_full).

mse_full_max :: Ensemble max of MSE(g_full, y_full).

mse_full_var :: Variance estimate of MSE(g_full, y_full), empirical (EKI/EKS) or quadrature (UKI).

mse_full_nn_mean :: MSE(g_full, y_full) of particle closest to the mean in parameter space. The mean in parameter space is the solution to the particle-based inversion.

failures :: Number of particle failures per iteration. If the calibration is run with the "high_loss" failure handler, this diagnostic will not capture the failures.

nn_mean_index :: Particle index of the nearest neighbor to the ensemble mean in parameter space.

CalibrateEDMF.Diagnostics.io_dictionary_particle_state — Function

io_dictionary_particle_state()

Dictionary of particle-wise parameter diagnostics, not involving forward model evaluations.

Elements:

u :: Parameters in unconstrained (inverse problem) space.

phi :: Parameters in constrained (physical) space.

CalibrateEDMF.Diagnostics.io_dictionary_particle_eval — Function

io_dictionary_particle_eval()

Dictionary of particle-wise parameter diagnostics involving forward model evaluations.

Elements:

g :: Forward model evaluation in inverse problem space.

g_full :: Forward model evaluation in primitive output space, normalized using the pooled field covariance.

mse_full :: Particle-wise evaluation of MSE(g_full, y_full).

CalibrateEDMF.Diagnostics.io_dictionary_prior — Function

io_dictionary_prior()

Parameter prior diagnostics dictionary.

Elements:

u_mean_prior :: Prior mean in unconstrained parameter space.

phi_mean_prior :: Prior mean in constrained parameter space.

u_var_prior :: Diagonal of the prior covariance in unconstrained space.

phi_low_unc_prior :: Lower uncertainty bound (μ-1σ_prior) of prior in constrained space.

phi_upp_unc_prior :: Upper uncertainty bound (μ+1σ_prior) of prior in constrained space.

phi_low_std_prior :: Lower standard bound (μ-1) of prior in constrained space. Useful measure of minimum allowed values for bounded parameters.

phi_upp_std_prior :: Upper standard bound (μ+1) of prior in constrained space. Useful measure of maximum allowed values for bounded parameters.

CalibrateEDMF.Diagnostics.get_ϕ_cov — Function

get_ϕ_cov(ekp::EnsembleKalmanProcess, priors::ParameterDistribution)

Get the last parameter covariance estimate in constrained (physical) space.

For ensemble methods, the covariance of the transformed parameters is returned. For unscented methods, the covariance is computed through a quadrature on the transformed quadrature points. The covariance of the transformed parameters returned here is equal to the transformed covariance only under a first order Taylor approximation, which is consistent with other approximations underlying the calibration method.

Inputs:

ekp :: The EnsembleKalmanProcess.

priors :: The priors defining transformations between constrained and unconstrained space.

Outputs:

The parameter covariance in constrained space.