ReferenceStats

CalibrateEDMF.ReferenceStats.ReferenceStatistics — Type

ReferenceStatistics{FT <: Real}

A structure containing statistics from the reference model used to define a well-posed inverse problem.

Fields

y::Vector{FT} where FT<:Real

Reference data, length: nSim * n_vars * n_zLevels (possibly reduced by PCA)

Γ::Matrix{FT} where FT<:Real

Data covariance matrix, dims: (y,y) (possibly reduced by PCA)

norm_vec::Array{Vector{FT}, 1} where FT<:Real

Vector (length: nSim) of normalizing factors (length: n_vars)

pca_vec::Array{Union{LinearAlgebra.UniformScaling, Matrix{FT}}, 1} where FT<:Real

Vector (length: nSim) of PCA projection matrices with leading eigenvectors as columns

y_full::Vector{FT} where FT<:Real

Full reference data vector, length: nSim * n_vars * n_zLevels

Γ_full::SparseArrays.SparseMatrixCSC{FT, Int64} where FT<:Real

Full covariance matrix, dims: (y,y)

Constructors

ReferenceStatistics(y::Vector{FT}, args...) where {FT <: Real}
ReferenceStatistics(
    RM::Vector{ReferenceModel};
    perform_PCA::Bool = true,
    normalize::Bool = true,
    variance_loss::FT = 0.1,
    tikhonov_noise::FT = 0.0,
    tikhonov_mode::String = "absolute",
    dim_scaling::Bool = false,
    y_type::ModelType = LES(),
    Σ_type::ModelType = LES(),
    Δt::FT = 6 * 3600.0,
) where {FT <: Real}

Constructs the ReferenceStatistics defining the inverse problem.

Inputs:

RM :: Vector of ReferenceModels.

perform_PCA :: Boolean specifying whether to perform PCA.

normalize :: Boolean specifying whether to normalize the data.

variance_loss :: Fraction of variance loss when performing PCA.

tikhonov_noise :: Tikhonov regularization factor for covariance matrices.

tikhonov_mode :: If "relative", tikhonov_noise is scaled by the maximum eigenvalue in the covariance matrix considered, having the interpretation of the inverse of the desired condition number. This value is enforced to be larger than the sqrt of the machine precision for stability.

dim_scaling :: Whether to scale covariance blocks by their size.

y_type :: Type of reference mean data. Either LES() or SCM().

Σ_type :: Type of reference covariance data. Either LES() or SCM().

Δt :: [LES last time - SCM start time (LES timeframe)] for LES_driven_SCM cases.

CalibrateEDMF.ReferenceStats.get_obs — Function

get_obs(
    m::ReferenceModel,
    y_names::Vector{String},
    Σ_names::Vector{String},
    normalize::Bool;
    z_scm::Union{Vector{FT}, Nothing} = nothing,
)
get_obs(
    m::ReferenceModel,
    y_type::Union{LES, SCM},
    Σ_type::Union{LES, SCM},
    normalize::Bool;
    z_scm::Union{Vector{FT}, Nothing},
)

Get observations for variables y_names, interpolated to z_scm (if given), and possibly normalized with respect to the pooled variance.

Inputs:

m :: Reference model

y_names :: Names of observed fields from the ReferenceModel m.

Σ_names :: Names of fields used to construct covariances, may be different than y_names if there are LES/SCM name discrepancies.

normalize :: Whether to normalize the observations.

z_scm :: If given, interpolate LES observations to given levels.

y_type :: ModelType used to construct observations, ::LES or ::SCM.

y_type :: ModelType used to construct noise covariances, ::LES or ::SCM.

Outputs:

y :: Mean of observations, possibly interpolated to z_scm levels.

Σ :: Observational covariance matrix, possibly pool-normalized.

pool_var :: Vector of vertically averaged time-variance, one entry for each variable

CalibrateEDMF.ReferenceStats.obs_PCA — Function

obs_PCA(y_mean, y_var, allowed_var_loss = 1.0e-1)

Perform dimensionality reduction using principal component analysis on the variance y_var. Only eigenvectors with eigenvalues that contribute to the leading 1-allowed_var_loss variance are retained. Inputs:

y_mean :: Mean of the observations.

y_var :: Variance of the observations.

allowed_var_loss :: Maximum variance loss allowed.

Outputs:

y_pca :: Projection of y_mean onto principal subspace spanned by eigenvectors.

y_var_pca :: Projection of y_var on principal subspace.

P_pca :: Projection matrix onto principal subspace, with leading eigenvectors as columns.

CalibrateEDMF.ReferenceStats.pca — Function

pca(covmat::AbstractMatrix{FT}, allowed_var_loss::FT) where {FT <: Real}

Perform dimensionality reduction using principal component analysis on the variance covmat.

Inputs:

covmat :: Variance of the observations.

allowed_var_loss :: Maximum variance loss allowed.

Outputs:

λ_pca :: Principal eigenvalues, ordered in increasing value order.

P_pca :: Projection matrix onto principal subspace, with leading eigenvectors as columns.

CalibrateEDMF.ReferenceStats.pca_length — Function

Returns dimensionality of the ReferenceStatistics in low-dimensional latent space

CalibrateEDMF.ReferenceStats.full_length — Function

Returns full dimensionality of the ReferenceStatistics, before latent space encoding

CalibrateEDMF.ReferenceStats.get_profile — Function

get_profile(
    filename::String,
    y_names::Vector{String};
    ti::Real = 0.0,
    tf = nothing,
    z_scm::Union{Vector{FT}, Nothing} = nothing,
)
get_profile(m::ReferenceModel, filename::String; z_scm::Union{Vector{T}, T} = nothing) where {T}
get_profile(
    m::ReferenceModel,
    filename::String,
    y_names::Vector{String};
    z_scm::Union{Vector{T}, T} = nothing,
) where {T}

Get time-averaged profiles for variables y_names, interpolated to z_scm (if given), and concatenated into a single output vector.

Inputs:

filename :: nc filename

y_names :: Names of variables to be retrieved.

ti :: Initial time of averaging window.

tf :: Final time of averaging window.

z_scm :: If given, interpolate LES observations to given levels.

m :: ReferenceModel from which to fetch profiles, implicitly defines ti and tf.

Outputs:

y :: Output vector used in the inverse problem, which concatenates the requested profiles.

CalibrateEDMF.ReferenceStats.get_time_covariance — Function

get_time_covariance(m::ReferenceModel, y_names::Vector{String}; z_scm::Vector{FT}) where {FT <: Real}

Obtain the covariance matrix of a group of profiles, where the covariance is obtained in time.

Inputs:

m :: Reference model.

y_names :: List of variable names to be included.

z_scm :: If given, interpolates covariance matrix to this locations.