TurbulenceConvectionUtils

CalibrateEDMF.TurbulenceConvectionUtils.ModelEvaluator — Type

ModelEvaluator

A structure containing the information required to perform a forward model evaluation at a given parameter vector.

Fields

param_cons::Vector{FT} where FT<:Real

Parameter vector in constrained (i.e. physical) space

param_names::Vector{String}

Parameter names associated with parameter vector

param_map::CalibrateEDMF.HelperFuncs.ParameterMap

A mapping operator to define relations between parameters. See ParameterMap for details.

ref_models::Vector{CalibrateEDMF.ReferenceModels.ReferenceModel}

Vector of reference models

ref_stats::CalibrateEDMF.ReferenceStats.ReferenceStatistics

Reference statistics for the inverse problem

Constructors

ModelEvaluator(param_cons, param_names, param_map, RM, RS)

defined at /home/runner/work/CalibrateEDMF.jl/CalibrateEDMF.jl/src/TurbulenceConvectionUtils.jl:63.

ModelEvaluator(; param_cons, param_names, param_map, ref_models, ref_stats)

CalibrateEDMF.TurbulenceConvectionUtils.run_SCM — Function

run_SCM(u, u_names, param_map, RM::Vector{ReferenceModel}, RS::ReferenceStatistics; 
        [error_check = false, failure_handler = "high_loss"])
run_SCM(ME::ModelEvaluator; [error_check = false, failure_handler = "high_loss"])

Run the single-column model (SCM) using a set of parameters and return the value of outputs defined in y_names, possibly after normalization and projection onto lower dimensional space using PCA.

The function also outputs a boolean diagnosing whether the requested SCM simulation failed.

Arguments

u :: Values of parameters to be used in simulations.

u_names :: SCM names for parameters u.

param_map :: A mapping operator to define relations between parameters. See also ParameterMap.

RM :: Vector of ReferenceModels

RS :: Reference statistics for simulation

Keywords

error_check :: Returns as an additional argument whether the SCM call errored.

failure_handler :: Method used to handle failed simulations.

Returns

sim_dirs :: Vector of simulation output directories

g_scm :: Vector of model evaluations concatenated for all flow configurations.

g_scm_pca :: Projection of g_scm onto principal subspace spanned by eigenvectors.

model_error :: Whether the simulation errored with the requested configuration.

CalibrateEDMF.TurbulenceConvectionUtils.run_reference_SCM — Function

run_reference_SCM(m::ReferenceModel; [overwrite = false, run_single_timestep = true])

Run the single-column model (SCM) for a reference model object using default parameters, and write the output to file.

Arguments:

m :: A ReferenceModel.

Keywords

output_root :: Directory within which the simulation results folder is saved to.

uuid :: Simulation identifier. Appended to the simulation results folder.

overwrite :: If true, run TC.jl and overwrite existing simulation files.

run_single_timestep :: If true, run only one time step.

CalibrateEDMF.TurbulenceConvectionUtils.eval_single_ref_model — Function

eval_single_ref_model(m_index, m::ReferenceModel, RS::ReferenceStatistics, u, u_names, param_map)

Run the single-column model (SCM) under a single configuration (i.e., ReferenceModel) using a list of parameters u with names u_names, and return the forward model evaluation in both the original and the latent PCA space.

Arguments

m_index :: Index of m in within the list of overarching ReferenceModels vector used to construct RS.

m :: A ReferenceModel.

RS :: The ReferenceStatistics for simulation.

u :: Values of parameters to be used in simulations.

u_names :: SCM names corresponding to parameter values u.

param_map :: A mapping operator to define relations between parameters. See also ParameterMap.

Returns

sim_dir::String :: Simulation output directory.

g_scm::Vector :: Forward model evaluation in original output space.

g_scm_pca::Vector :: Projection of g_scm onto principal subspace spanned by eigenvectors.

model_error::Bool :: Whether the simulation errored with the requested configuration.

CalibrateEDMF.TurbulenceConvectionUtils.run_SCM_handler — Function

run_SCM_handler(case_name::String, out_dir, u, u_names, param_map; [namelist, namelist_args, uuid, les])
run_SCM_handler(m::ReferenceModel, out_dir, u, u_names, param_map)

Run a TurbulenceConvection.jl case. Return directory path where simulation data is stored and a flag for whether the simulation failed.

The case is specified with the first function argument, either by providing the case name as a String, or by passing a ReferenceModel m. The out_dir is the path where the output directory is to be stored. The returned path is therefore of the form out_dir/Output.case_name.uuid where uuid is a simulation run identifier (defaults to "01").

The keywords u and u_names are vectors of parameter values and names that will be modified in the TC.jl run.

If a case_name is provided, a namelist can optionally be provided that defines all case-specific parameters and run-options needed to run TC.jl. If this not provided, the default case-specific namelist from TC.jl is used. Additional arguments to modify the TC.jl namelist can also be provided to namelist_args; see also change_entry!.

The ParameterMap param_map sets the values of specific parameters provided pre-defined mapping operators. See also ParameterMap.

If the case is LES_driven_SCM, a path to an LES stats file must be provided to les. See also get_cfsite_les_dir.

Note that if a ReferenceModel m is passed, namelist, uuid and les are not valid keywords. In this case, namelist and les are fetched from the ReferenceModel, and uuid is a random string.

Arguments

m :: A ReferenceModel

case_name :: Name of case to run. For example: Bomex, TRMM_LBA, LES_driven_SCM.

out_dir :: Path to where folder with simulation results are to be saved.

u :: Values of parameters to be used in simulations.

u_names :: SCM names for parameters u.

param_map :: A mapping operator to define relations between parameters. See ParameterMap for details.

Keywords

namelist :: namelist to use for simulation.

uuid :: uuid of SCM run

les :: path to LES stats file, or NamedTuple with keywords {forcingmodel, month, experiment, cfsitenumber} needed to specify path.

Returns

String :: Directory path where output data from the TC.jl run is stored.

Bool :: true if the simulation failed/crashed, false otherwise.

CalibrateEDMF.TurbulenceConvectionUtils.create_parameter_vectors — Function

create_parameter_vectors(u_names, u, param_map, namelist)

Given vector of parameter names and corresponding values, combine any vector components into single parameter vectors for input into SCM.

Arguments:

u_names :: SCM names for parameters u, which may contain vector components.

u :: Values of parameters to be used in simulations, which may contain vector components.

param_map :: A mapping to a reduced parameter set. See ParameterMap for details.

namelist :: The parameter namelist for TurbulenceConvection.jl

Returns:

Vector{String} :: SCM names for parameters u.

Vector{Vector{Real}} :: Values of parameters to be used in simulations.

CalibrateEDMF.TurbulenceConvectionUtils.generate_scm_input — Function

generate_scm_input(model_evaluators, iteration, [outdir_path = pwd(), batch_indices = nothing])

Writes to file a set of ModelEvaluator used to initialize SCM evaluations at different parameter vectors, as well as their assigned version, which is on the form i<iteration>_e<ensemble_index>.

CalibrateEDMF.TurbulenceConvectionUtils.parse_version_inds — Function

parse_version_inds(version)

Given version = "ix_ey", return the iteration index x and ensemble index y as integers.

CalibrateEDMF.TurbulenceConvectionUtils.precondition — Function

precondition(param::Vector{<:Real}, priors, param_map, ref_models, ref_stats; [counter, max_counter = 10])

Substitute parameter vector param by a parameter vector drawn from the same prior, conditioned on the forward model being stable.

Arguments

param :: A parameter vector that may possibly result in unstable forward model evaluations (in unconstrained space).

priors :: Priors from which the parameters were drawn.

param_map :: A mapping to a reduced parameter set. See ParameterMap for details.

ref_models :: Vector of ReferenceModels to check stability for.

ref_stats :: ReferenceStatistics of the ReferenceModels.

counter :: Accumulator tracking number of recursive calls to preconditioner.

max_counter :: Maximum number of recursive calls to the preconditioner.

Returns

Vector{<:Real} :: A new parameter vector drawn from the prior, conditioned on simulations being stable (in unconstrained space).

precondition(ME::ModelEvaluator, priors)

Substitute the parameter vector of a ModelEvaluator by another one drawn from the given priors, conditioned on the forward model being stable.

Arguments

ME :: A ModelEvaluator.

priors :: Priors from which the parameters were drawn.

Returns

ModelEvaluator :: A preconditioned ModelEvaluator.

CalibrateEDMF.TurbulenceConvectionUtils.get_gcm_les_uuid — Function

get_gcm_les_uuid(cfsite_number; [forcing_model::String, month, experiment])

Generate unique and self-describing uuid given information about a GCM-driven LES simulation from Zhaoyi Shen, Akshay Sridhar, Zhihong Tan, Anna Jaruga, Tapio Schneider (2022).

Examples

julia> get_gcm_les_uuid(1; forcing_model = "HadGEM2-A", month = 7, experiment = "amip")
"1_HadGEM2-A_07_amip"

CalibrateEDMF.TurbulenceConvectionUtils.save_tc_data — Function

save_tc_data(cases, outdir_path, sim_dirs, version, suffix)

Save full TC.jl output in <results_folder>/timeseries.<suffix>/iter_<iteration>/Output.<case>.<case_id>_<ens_i>.

If the flag save_tc_output is set to true in the output config, i.e. config["output"], TC.jl output is saved, and if additionally a list of iterations is specified in save_tc_iterations (also in the output config), only those EKP iterations are saved.

Arguments

config :: The calibration config dictionary.

outdir_path :: The results results_folder path

sim_dirs :: List of (temporary) directories where raw simulation output is initially saved to

version :: An identifier for the current iteration and ensemble index

suffix :: Case set identifier; is either "train" or "validation".